whisper-server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
torch>=2.0.0
requests>=2.31.0
